fix(genesis,params): moving genesis params to genesis module
