+++ genesis/genesis.go
@@ -11,7 +11,6 @@ import (
 	"github.com/pactus-project/pactus/crypto/hash"
 	"github.com/pactus-project/pactus/types/account"
 	"github.com/pactus-project/pactus/types/amount"
-	"github.com/pactus-project/pactus/types/param"
 	"github.com/pactus-project/pactus/types/validator"
 	"github.com/pactus-project/pactus/util"
 )
@@ -56,10 +55,10 @@ type Genesis struct {
 }
 
 type genesisData struct {
-	GenesisTime time.Time        `cbor:"1,keyasint" json:"genesis_time"`
-	Params      *param.GenParams `cbor:"2,keyasint" json:"params"`
-	Accounts    []genAccount     `cbor:"3,keyasint" json:"accounts"`
-	Validators  []genValidator   `cbor:"4,keyasint" json:"validators"`
+	GenesisTime time.Time      `cbor:"1,keyasint" json:"genesis_time"`
+	Params      *GenParams     `cbor:"2,keyasint" json:"params"`
+	Accounts    []genAccount   `cbor:"3,keyasint" json:"accounts"`
+	Validators  []genValidator `cbor:"4,keyasint" json:"validators"`
 }
 
 func (gen *Genesis) Hash() hash.Hash {
@@ -72,7 +71,7 @@ func (gen *Genesis) GenesisTime() time.Time {
 	return gen.data.GenesisTime
 }
 
-func (gen *Genesis) Params() *param.GenParams {
+func (gen *Genesis) Params() *GenParams {
 	return gen.data.Params
 }
 
@@ -124,7 +123,7 @@ func makeGenesisValidator(val *validator.Validator) genValidator {
 }
 
 func MakeGenesis(genesisTime time.Time, accounts map[crypto.Address]*account.Account,
-	validators []*validator.Validator, params *param.GenParams,
+	validators []*validator.Validator, params *GenParams,
 ) *Genesis {
 	genAccs := make([]genAccount, len(accounts))
 	for addr, acc := range accounts {
