+++ genesis/genesis_test.go
@@ -10,7 +10,6 @@ import (
 	"github.com/pactus-project/pactus/genesis"
 	"github.com/pactus-project/pactus/types/account"
 	"github.com/pactus-project/pactus/types/amount"
-	"github.com/pactus-project/pactus/types/param"
 	"github.com/pactus-project/pactus/types/validator"
 	"github.com/pactus-project/pactus/util"
 	"github.com/pactus-project/pactus/util/testsuite"
@@ -26,7 +25,7 @@ func TestMarshaling(t *testing.T) {
 	val, _ := ts.GenerateTestValidator(0)
 	gen1 := genesis.MakeGenesis(util.RoundNow(10),
 		map[crypto.Address]*account.Account{prv: acc},
-		[]*validator.Validator{val}, param.DefaultGenParams())
+		[]*validator.Validator{val}, genesis.DefaultGenParams())
 	gen2 := new(genesis.Genesis)
 
 	assert.Equal(t, 10, gen1.Params().BlockIntervalInSecond)
@@ -95,7 +94,7 @@ func TestCheckGenesisAccountAndValidator(t *testing.T) {
 		accs[pub.AccountAddress()] = acc
 		vals = append(vals, val)
 	}
-	gen := genesis.MakeGenesis(time.Now(), accs, vals, param.DefaultGenParams())
+	gen := genesis.MakeGenesis(time.Now(), accs, vals, genesis.DefaultGenParams())
 
 	for addr, acc := range gen.Accounts() {
 		assert.Equal(t, accs[addr], acc)
